% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicates.R
\docType{data}
\name{replicates}
\alias{replicates}
\title{Raw mass spectrum proteomics log abundance for 4 pairs of technical replicates.}
\format{A data frame of 85 rows and 8 columns with missing peaks' abundance as NA.}
\usage{
replicates
}
\description{
Raw mass spectrum proteomics log abundance for 4 pairs of technical replicates.
}
\keyword{datasets}
