% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.bc.full.R
\docType{data}
\name{tcga.bc.full}
\alias{tcga.bc.full}
\title{A subset of mass spectrum metabolomics data for TCGA breast cancer study without missing peaks.}
\format{A data frame of 100 rows and 30 columns without missing value (NA).}
\usage{
tcga.bc.full
}
\description{
A subset of mass spectrum metabolomics data for TCGA breast cancer study without missing peaks.
}
\keyword{datasets}
