% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.bc.R
\docType{data}
\name{tcga.bc}
\alias{tcga.bc}
\title{Raw mass spectrum metabolomics data for TCGA breast cancer study.}
\format{A data frame of 40 rows and 30 columns with missing peaks' abundance as NA.}
\usage{
tcga.bc
}
\description{
Raw mass spectrum metabolomics data for TCGA breast cancer study.
}
\keyword{datasets}
